requirements.txt
Django==3.1.12
djongo==1.3.7
pymongo<=3.11.4
celery==5.3.0
redis==4.5.0
requests==2.31.0
python-dotenv==1.0.0
python-telegram-bot==13.7

